VECTOR3: don't rely on UB
Accessing both members of a union at the same time is an undefined
behavior in c++.

VECTOR3 does already provide operator[] to access its members by
index, so instead of this operator relying on the "data" member
variable, we create one on the fly. All major compilers (gcc, clang and
msvc) optimize it away with -O2, so there is no performance penalty.

All direct accesses to "data" are also replaced with the index operator.