Change 32-bit builds to be large-address aware (#542)
This allows libraries/addons with large amounts of textures to avoid erroneous out-of-memory texture loading failures, which happens with NASSP for unknown reasons with some graphics cards. We have done a lot of testing and cannot find any indication that the process is actually consuming up to its maximum RAM limit, but nevertheless by extending the limit from 2GB to 4GB with this linker flag, the error is resolved.